Move SystemEventStream to EventStreamAdapter
